+++ include/ck/tensor_operation/gpu/device/impl/device_grouped_gemm_multiple_d_dl.hpp
@@ -607,10 +607,10 @@ struct DeviceGroupedGemmMultipleD_Dl : public DeviceGroupedGemm<ALayout,
                 }
             }
 
-            if(cpy_stream &&
-               cpy_event) // If user provides copy stream and copy event, we assume that he's also
-                          // responsible for providing allocated host memory (eg. pinned) which
-                          // would be used to copy kernel arguments to the device.
+            // If user provides copy stream and copy event, we assume that he's also
+            // responsible for providing allocated host memory (eg. pinned) which
+            // would be used to copy kernel arguments to the device.
+            if(cpy_stream && cpy_event)
             {
                 if(arg.gemm_kernel_host_args_ == nullptr)
                 {
@@ -810,9 +810,8 @@ struct DeviceGroupedGemmMultipleD_Dl : public DeviceGroupedGemm<ALayout,
     ///             achieve fully async data copy.
     ///
     /// @param      p_arg              The pointer to the Argument we're going to update.
-    /// @param[in]  p_host_kernel_args The pointer to the host memory which contains kernel
-    ///                                arguments.
-    ///
+    /// @param[in]  p_host_kernel_args The pointer to the host memory where the kernel
+    ///                                arguments will be copied
     void SetHostKernelArgsPointer(BaseArgument* p_arg, void* p_host_kernel_args) const
     {
         Argument* pArg_ = dynamic_cast<Argument*>(p_arg);